アプリの構成を変更
