+++ app/src/main/java/com/android/example/tasklist/Controller/Activity/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.android.example.tasklist
+package com.android.example.tasklist.Controller.Activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
