+++ app/src/main/java/com/android/example/tasklist/Controller/Fragment/FavoriteListFragment.kt
@@ -1,12 +1,13 @@
-package com.android.example.tasklist
+package com.android.example.tasklist.Controller.Fragment
 
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.android.example.tasklist.Model.Task
+import com.android.example.tasklist.View.TaskListAdapter
 import com.android.example.tasklist.databinding.FragmentFavoriteListBinding
 
 class FavoriteListFragment : Fragment() {
