+++ app/src/main/java/com/android/example/tasklist/Controller/Fragment/TaskListFragment.kt
@@ -1,4 +1,4 @@
-package com.android.example.tasklist
+package com.android.example.tasklist.Controller.Fragment
 
 import android.annotation.SuppressLint
 import android.content.Context
@@ -13,10 +13,13 @@ import android.view.ViewGroup
 import android.widget.ImageButton
 import android.widget.Toast
 import androidx.core.content.edit
-import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.setFragmentResult
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.android.example.tasklist.Model.MstColor
+import com.android.example.tasklist.R
+import com.android.example.tasklist.Model.Task
+import com.android.example.tasklist.View.TaskListAdapter
 import com.android.example.tasklist.databinding.FragmentTaskListBinding
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
@@ -56,7 +59,7 @@ class TaskListFragment : Fragment() {
             }
         }
 
-        adapter.setOnItemClickListener(object:TaskListAdapter.OnItemClickListener{
+        adapter.setOnItemClickListener(object: TaskListAdapter.OnItemClickListener {
             override fun onItemClickListener(view: View, position: Int, clickedText: Task) {
                 val button = view.findViewById<ImageButton>(R.id.add_Favorite)
