+++ app/src/main/res/layout/fragment_add_task.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".AddTaskFragment">
+    tools:context=".Controller.Fragment.AddTaskFragment">
 
     <TextView
         android:id="@+id/title_Label"
