+++ app/src/main/res/layout/fragment_favorite_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FavoriteListFragment" >
+    tools:context=".Controller.Fragment.FavoriteListFragment" >
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerView"
