+++ app/src/main/res/layout/fragment_main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainFragment">
+    tools:context=".Controller.Fragment.MainFragment">
 
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/viewPager"
